+++ cdoc2-lib/src/test/java/ee/cyber/cdoc2/smartid/JWSSignerTest.java
@@ -98,7 +98,7 @@ class JWSSignerTest {
         final String[] verificationCode = {null};
 
         InteractionParams interactionParams = InteractionParams
-            .displayTextAndVCCForDocument("JWSSignerTest::testSignature")
+            .displayTextAndVCCForDocument("JWSSignerTest::testSignature", null, null)
             .addAuthListener(e -> {
                 verificationCode[0] = e.getVerificationCode();
                 log.debug("Verification code: {}", verificationCode[0]);